+++ NEWS.md
@@ -1,4 +1,4 @@
-# Version (development version)
+# Version 1.35.0 [2023-03-22]
 
 ## New Features